refactor(src/components): Create src and moved sub-folder
