+++ src/components/hero.tsx
