+++ src/components/submit-button.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Button } from "@/components/ui/button";
+import { Button } from "@/src/components/ui/button";
 import { type ComponentProps } from "react";
 import { useFormStatus } from "react-dom";
